requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
pytz==2022.7.1
